Remove CreateView from base Plugin class
CreateView needs to know about the current UI type, QWidgets in this case.
Since all calls to this function are made from the GUI, it is moved to somewhere in that folder.
There are at the moment no calls to createView in effect.
For the other two, instrument and tool, a new class has been created, to separate QWidget
plugins from any other future kind. In that new class is where the createView is redefined.