Add QML GUI implementation Proof of Concept
This commit removes compilation of QWidgets GUI and targets Qt6.6.
These changes should be done under a compile flag in the future so both
versions can still compile for a while.

The QML application barely does anything other than starting the core
LMMS, loading a project and crashing because of a missing pluing.

When not loading any plugin, there is a minimal interface that resembles the
QML design posted in the forums.