Android build prototype
In order to build a working version the following libraries need to be
cross compiled and present:
* fftw 3.3.10
* libsamplerate 0.2.2
* libsndfile 1.2.2
* qt 6.7

As a manual step, plugin libraries need to be copied to the
android/libs/arch folder before running.

The missing functionality compared to linux build is
* no VST support (this will never work on Android)
* AudioFileProcessor doesn't load samples properly

Because of file system permissions in newer android, before loading,
all the resources will be copied to a folder in the application folder
by the android activity.